Started breaking up the HiGlassComponent tests
