+++ karma.conf.js
@@ -43,7 +43,8 @@ module.exports = (config) => {
       // 'test/GenomePositionSearchBoxTest.js',
       // 'test/HeatmapTests.js',
       // 'test/HiGlassComponentCreationTests.js',
-      'test/HiGlassComponentTest.js',
+      // 'test/HiGlassComponentTest.js',
+      'test/HiGlassComponent/*.js',
       // 'test/Horizontal1DTrackTests.js',
       // 'test/HorizontalHeatmapTests.js',
       // 'test/HorizontalMultivecTests.js',
